[TASK] Introducing traits for holding the type properties
A type can inherit from multiple other types. The properties of a type
are defined in a trait (if own properties are available). Each model
uses the traits to include the properties.

A type view helper is now only an envelope for a specific type. The logic
is done in the Abstract view helper class which uses the according
model to register its arguments.

A shell script is now used to generate the type models and view helpers
and format them according to the code styles.