[FEATURE] Specific type of WebPage can be selected in the page properties
A new page field "Specific type of web page" was introduced (WebPage, AboutPage, ItemPage, CollectionPage, ...)
to specify the type of page. Also an extension configuration setting is now available to activate or deactivate
the automatic injection of the WebPage schema into the page, when at page generation time none is specified.

The WebPage type can be set via API:

```
$webPage = (new \Brotkrueml\Schema\Model\Type\WebPage())
    ->setProperty('name', 'The title of the web page')
    ->setProperty('description', 'The description of the web page')
    ->setProperty('primaryImageOfPage', 'https://example.org/image.png')
;

$schemaManager->setWebPage($webPage);
```

or a view helper:
```
<schema:type.webPage
    name="The title of the web page"
    description="The description of the web page"
    primaryImageOfPage="https://example.org/image.png"
/>
```

If these are called more the one, the last one wins - there can only be ony WebPage on a page.

Resolves: #1