[FEATURE] BreadcrumbLists can be handled by SchemaManager
BreadcrumbLists can be added to the Schema Manager. When rendering the JSON-LD result the BreadcrumbLists are assigned to the WebPage.

A new view helper <schema:breadcrumb> was implemented to ease the assignment of a breadcrumb zu the page. The view helper accepts the output of the MenuProcessor for building a BreadcrumbList.

The Schema Manager API for adding a web page was streamlined, now a WebPage can be added with the ->addType() method. The ->setWebPage() method was removed.

Resolves: #2