[TASK] Catch exception when retrieving CacheManager
This is a preparation for compatibility with TYPO3 v12. The exception
"TYPO3\CMS\Core\Cache\CacheManager can not be injected/instantiated during
ext_localconf.php or TCA loading. Use lazy loading instead." is thrown.

This is due to the initialisation of the page field "tx_schema_webpagetype" in
TCA which uses a class method to fill the items array - which itself gets the
items from a cache at the end. This is not needed at build time and the
exception can be ignored safely.

See also: https://github.com/TYPO3/typo3/commit/a92d9430c0e1daeb25d8129c74ca18ee2ed9cb92