refactor: use attribute to determine type
With #107 registering types via the PHP attribute `Brotkrueml\Schema\Attributes\Type`
was introduced to decouple the class name from the type. However, the
AbstractType->getType() method still checks for the class name (and has
a logic when a type starts with a number and therefore the class starts with
an underscore).

Now the attribute is taken into account. To speed up the determination of the type,
a static property is introduced which caches already resolved types.