refactor: dissolve internal InitialiseTypesEvent
The SchemaManager is cleaned up: The previous InitialiseTypesEvent which was fired for
adding the initial BreadcumbList when the corresponding extension setting is activated
has been dropped. The existing RenderAdditionalTypesEvent is used for that case now.

This has the impact that the automatically generated BreadcrumbList is not at the
end of the breadcrumb list, not at the beginning like before (when having more
than one BreadcrumbList). As it is a list, the ordering should not matter.