fix: avoid call to TSFE->get_cache_timeout()
In TYPO3 v13 the `get_cache_timeout()` method is marked as protected.
The default lifetime of a page is 24 hours, so this is now set.
If the cache is cleared at midnight, the cache entry is also cleared
(because of the cache tag). So, this settings should do no harm.