chore!: drop extension setting "Embed markup in the body section"
Historically, the extension added the schema markup in the head section. Later on,
a setting was introduced to let the integrator decide, if it should be added to the
head or the body. However, search engines don't care about the position. It might be
better to move the markup away from the head to speed up parsing of HTML for browsers.

See: https://youtu.be/lI6EtxjoyDU