Consider type safety of view helper arguments in Fluid 5
Fluid 5 (shipped with TYPO3 v14) introduced type safety for view helper arguments.
This means, that a string `true` is now casted automatically to an int for the
`isMainEntityOfWebPage` view helper argument, resulting in `0`.

When we are in TYPO3 v14, there is now a check and the type of that
argument is adjusted accordingly. Additionally, some tests have been
added to test against the boolean variable `{true}`, introduced with
Fluid 4 (and shipped with TYPO3 v13).

The argument can have three states:
- 0 = not a main entity
- 1 = add to main entities
- 2 = prioritise as main entity (move already assigned main entities to default types)

This way, we have to use an int for now.